--- service/auth/auth_dto/src/main/java/cafe/auth_dto/AuthDtoApplication.java
@@ -1,13 +1,0 @@
-package cafe.auth_dto;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AuthDtoApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AuthDtoApplication.class, args);
-    }
-
-}
