+++ service/auth/server/build.gradle
@@ -25,7 +25,7 @@ repositories {
 
 dependencies {
     implementation project(':common:domain')
-//    implementation project(':service:user:user_dto')
+    implementation project(':service:user:user_dto')
     implementation project(':service:auth:auth_dto')
 
     implementation 'org.springframework.boot:spring-boot-starter'
