+++ service/auth/server/src/main/java/cafe/auth/server/application/dto/AuthResponse.java
@@ -1,0 +1,22 @@
+package cafe.auth.server.application.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+public class AuthResponse {
+
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    @Builder
+    public static class SignIn {
+
+        private String token;
+
+        public static SignIn of(String token) {
+            return SignIn.builder().token(token).build();
+        }
+    }
+}
