+++ service/auth/server/src/main/java/cafe/auth/server/domain/JwtConstant.java
@@ -1,0 +1,12 @@
+package cafe.auth.server.domain;
+
+public interface JwtConstant {
+
+    String USER_ID = "USER_ID";
+
+    String USER_NAME = "USER_NAME";
+
+    String USER_ROLE = "USER_ROLE";
+
+    Long MILLI_SECOND = 1000L;
+}
