+++ service/auth/server/src/main/java/cafe/auth/server/exception/AuthErrorCode.java
@@ -1,0 +1,18 @@
+package cafe.auth.server.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum AuthErrorCode {
+
+    SIGN_IN_FAIL(HttpStatus.UNAUTHORIZED, "로그인 실패."),
+    TOKEN_EXPIRED(HttpStatus.UNAUTHORIZED, "만료된 토큰입니다."),
+    INVALID_TOKEN(HttpStatus.UNAUTHORIZED, "잘못된 토큰입니다."),
+    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "내부 서버 오류");
+
+    private final HttpStatus status;
+    private final String message;
+}
