+++ service/auth/server/src/main/java/cafe/auth/server/exception/AuthException.java
@@ -1,0 +1,15 @@
+package cafe.auth.server.exception;
+
+import cafe.domain.exception.BusinessException;
+import lombok.Getter;
+
+@Getter
+public class AuthException extends BusinessException {
+
+  AuthErrorCode errorCode;
+
+  public AuthException(AuthErrorCode errorCode) {
+    super(errorCode.getStatus().name(), errorCode.getMessage());
+    this.errorCode = errorCode;
+  }
+}
