+++ service/auth/server/src/main/java/cafe/auth/server/infrastructure/properties/JwtProperties.java
@@ -1,0 +1,16 @@
+package cafe.auth.server.infrastructure.properties;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Getter
+@Setter
+@Component
+@ConfigurationProperties(prefix = "jwt")
+public class JwtProperties {
+
+    private String secretKey;
+    private int accessTokenExpireIn;
+}
