+++ service/auth/server/src/main/java/cafe/auth/server/presentation/request/AuthRequest.java
@@ -1,0 +1,16 @@
+package cafe.auth.server.presentation.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+public class AuthRequest {
+
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class SignIn {
+        private String username;
+        private String password;
+    }
+}
