--- service/auth/server/src/main/java/cafe/server/AuthApplication.java
@@ -1,13 +1,0 @@
-package cafe.server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-
-@SpringBootApplication(exclude = { DataSourceAutoConfiguration.class })
-public class AuthApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AuthApplication.class, args);
-    }
-}
