--- service/auth/server/src/main/java/cafe/server/application/dto/AuthResponse.java
@@ -1,22 +1,0 @@
-package cafe.server.application.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-public class AuthResponse {
-
-    @Getter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    @Builder
-    public static class SignIn {
-
-        private String token;
-
-        public static SignIn of(String token) {
-            return SignIn.builder().token(token).build();
-        }
-    }
-}
