--- service/auth/server/src/main/java/cafe/server/domain/JwtConstant.java
@@ -1,12 +1,0 @@
-package cafe.server.domain;
-
-public interface JwtConstant {
-
-    String USER_ID = "USER_ID";
-
-    String USER_NAME = "USER_NAME";
-
-    String USER_ROLE = "USER_ROLE";
-
-    Long MILLI_SECOND = 1000L;
-}
