--- service/auth/server/src/main/java/cafe/server/exception/AuthException.java
@@ -1,15 +1,0 @@
-package cafe.server.exception;
-
-import cafe.domain.exception.BusinessException;
-import lombok.Getter;
-
-@Getter
-public class AuthException extends BusinessException {
-
-  AuthErrorCode errorCode;
-
-  public AuthException(AuthErrorCode errorCode) {
-    super(errorCode.getStatus().name(), errorCode.getMessage());
-    this.errorCode = errorCode;
-  }
-}
