--- service/auth/server/src/main/java/cafe/server/presentation/AuthController.java
@@ -1,19 +1,0 @@
-package cafe.server.presentation;
-
-import cafe.server.application.service.AuthService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@RequestMapping("/api/auth")
-@RestController
-public class AuthController {
-
-    private final AuthService authService;
-
-//    @PostMapping("/sign-in")
-//    public ApiRes
-
-}
