+++ service/auth/server/src/test/java/cafe/auth/server/AuthApplicationTests.java
@@ -1,0 +1,7 @@
+package cafe.auth.server;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class AuthApplicationTests {
+}
