--- service/auth/server/src/test/java/cafe/server/AuthApplicationTests.java
@@ -1,7 +1,0 @@
-package cafe.server;
-
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class AuthApplicationTests {
-}
