+++ service/eureka/server/src/test/java/cafe/auth/server/EurekaApplicationTests.java
@@ -1,0 +1,14 @@
+package cafe.auth.server;
+
+import cafe.eureka.EurekaApplication;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(classes = EurekaApplication.class)
+class EurekaApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
