--- service/eureka/server/src/test/java/cafe/server/EurekaApplicationTests.java
@@ -1,14 +1,0 @@
-package cafe.server;
-
-import cafe.eureka.EurekaApplication;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest(classes = EurekaApplication.class)
-class EurekaApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
