+++ service/gateway/server/src/main/java/cafe/gateway/application/AuthService.java
@@ -1,6 +1,6 @@
 package cafe.gateway.application;
 
-import cafe.auth_dto.jwt.JwtClaim;
+import cafe.auth.auth_dto.jwt.JwtClaim;
 
 public interface AuthService {
