+++ service/gateway/server/src/test/java/cafe/auth/server/GatewayApplicationTests.java
@@ -1,0 +1,14 @@
+package cafe.auth.server;
+
+import cafe.gateway.GatewayApplication;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(classes = GatewayApplication.class)
+class GatewayApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
