--- service/gateway/server/src/test/java/cafe/server/GatewayApplicationTests.java
@@ -1,14 +1,0 @@
-package cafe.server;
-
-import cafe.gateway.GatewayApplication;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest(classes = GatewayApplication.class)
-class GatewayApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
