+++ service/user/user_dto/src/main/java/cafe.user.user_dto/UserDtoApplication.java
@@ -1,0 +1,13 @@
+package cafe.user.user_dto;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserDtoApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(UserDtoApplication.class, args);
+    }
+
+}
